requirements.txt
pydantic~=2.10.5
tqdm~=4.67.1
python-dotenv~=1.0.1
pandas~=2.2.3
requests~=2.32.3
nltk~=3.9.1
tenacity~=9.0.0
datasets~=3.2.0
openai~=1.59.7
groq~=0.18.0
httpx~=0.28.1
numpy~=2.2.1
matplotlib~=3.10.0
seaborn~=0.13.2
statsmodels~=0.14.4
yfinance~=0.2.51
certifi~=2024.12.14
scipy~=1.15.1
scikit-learn~=1.6.1
beautifulsoup4~=4.12.3
torch~=2.6.0
transformers~=4.48.2
openpyxl~=3.1.5
aiohttp~=3.11.11